[kda合入] kda版本纯增量代码合入(constant, softplus, index, kda版本cumsum)
