+++ fla/ops/triton/triton_core/kda/chunk_bwd.py
@@ -19,9 +19,9 @@ from fla.ops.triton.triton_core.kda._kda_cp.chunk_delta_h import (
 from fla.ops.triton.triton_core.kda.chunk_intra import chunk_kda_bwd_intra
 from fla.ops.triton.triton_core.kda.gate import kda_gate_bwd, kda_gate_chunk_cumsum
 from fla.ops.triton.triton_core.kda.wy_fast import recompute_w_u_fwd
-from fla.ops.triton.triton_core.kda._kda_utils.cumsum import chunk_local_cumsum
-from fla.ops.triton.triton_core.kda._kda_utils.index import prepare_chunk_indices
-from fla.ops.triton.triton_core.kda._kda_utils.constant import RCP_LN2
+from fla.ops.triton.triton_core.cumsum import chunk_local_cumsum
+from fla.ops.triton.triton_core.index import prepare_chunk_indices
+from fla.ops.triton.triton_core.constant import RCP_LN2
 from fla.ops.triton.triton_core.kda._kda_utils.op import exp2
 from fla.ops.triton.triton_core.kda._kda_utils.utils import (
     autotune_cache_kwargs,
@@ -853,6 +853,7 @@ def chunk_kda_bwd(
         reverse=True,
         cu_seqlens=cu_seqlens,
         chunk_indices=chunk_indices,
+        use_per_head_kernel=True,
     )
     if use_gate_in_kernel:
         dg, dA, dbias = kda_gate_bwd(
