+++ fla/ops/triton/triton_core/kda/chunk_intra.py
@@ -12,7 +12,7 @@ import triton.language as tl
 
 from fla.ops.triton.triton_core.kda.chunk_intra_token_parallel import chunk_kda_fwd_intra_token_parallel
 from fla.ops.triton.triton_core.kda.wy_fast import recompute_w_u_fwd
-from fla.ops.triton.triton_core.kda._kda_utils.index import prepare_chunk_indices
+from fla.ops.triton.triton_core.index import prepare_chunk_indices
 from fla.ops.triton.triton_core.kda._kda_utils.op import exp2, gather
 from fla.ops.triton.triton_core.kda._kda_utils.utils import IS_GATHER_SUPPORTED, autotune_cache_kwargs
